requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.7
rq==1.16.2
email-validator==2.2.0
dnspython==2.6.1
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.0.3
